types: unsafe_subst -> traverse
Also makes unsafe_subst less generic. The addition of a
`Var` constructor was considered, but rejected because of the
low benefits (this is a rare case where we need to take either
a bound or a free var) compared to the overhead of nesting the
types in `term` (flat is better).